LayoutTests/imported/w3c/web-platform-tests/fetch/security/dangling-markup/textarea-expected.txt



FAIL       <form action="/resource-timing/resources/document-navigated.html" method="post">        <input type="submit">        <textarea name="dangling">     assert_equals: expected "error" but got "submit"
FAIL       <div>        <form action="/resource-timing/resources/document-navigated.html" method="post">          <input type="submit">          <textarea name="dangling">     assert_equals: expected "error" but got "submit"
FAIL       <form action="/resource-timing/resources/document-navigated.html" method="post" id="form">        <input type="submit">      </form>      <textarea name="dangling" form="form">     assert_equals: expected "error" but got "submit"
